+++ Survey2/Survey_2/Indexing/Index.java
@@ -1,0 +1,3 @@
+public interface Index{
+    public float calculate(float pix1, float pix2);
+}