+++ plugins/Calibration/Calibrator.java
@@ -47,6 +47,10 @@ public class Calibrator{
   public static String MAP_TARG2 = "TARGET2";
   public static String MAP_TARG3 = "TARGET3";
 
+  private Debugger debugger = Debugger.getInstance(false);
+  private boolean ReflectanceImageOnly = false;
+
+
 
 
   public ImagePlus[] scaleImages(ImagePlus[] channels){
@@ -236,10 +240,6 @@ public class Calibrator{
     double[] regressionParams = null;
     double r_Squared = 0.0;
 
-    /* For experimenting
-    for( int i=0; i<calcValues.length; i++ ){
-      calcValues[i] = (double)calcValues[i]/100.0;
-    }*/
 
     CurveFitter visRegression = new CurveFitter(calcValues, refValues);
     visRegression.doFit(0, true);
@@ -255,7 +255,7 @@ public class Calibrator{
     PlotWindow.noGridLines = false;
     Plot visPlot = new Plot("Visible band regression", "Image values", "Reflectance values");
     //visPlot.setLimits(10.0/100f, 30.0/100f, 0.0, 1.0);
-    visPlot.setLimits(0.0, 1.0, 0.0, 1.0);
+    visPlot.setLimits(0.0, 255.0, 0.0, 1.0);
     visPlot.setColor(Color.RED);
     visPlot.addPoints(calcValues, refValues, 0);
     visPlot.draw();
@@ -263,7 +263,16 @@ public class Calibrator{
     double[] yVis = new double[]{regressionParams[0], regressionParams[1] + regressionParams[0]};
     visPlot.addPoints(xVis, yVis, 2);
     visPlot.addLabel(0.05, 0.1, "R squared = " + Double.toString(r_Squared));
-    visPlot.show();
+
+    if( Debugger.DEBUGMODE ){
+        for( int i=0; i<calcValues.length; i++ ){
+          //calcValues[i] = (double)calcValues[i]/100.0;
+          IJ.log( (String)"Target " + (i+1) + " Mean: " + calcValues[i]);
+        }
+        IJ.log((String)("intercept: " + IJ.d2s((double)regressionParams[0], (int)8)));
+        IJ.log((String)("slope: " + IJ.d2s((double)regressionParams[1], (int)8)));
+        visPlot.show();
+    }
 
     //HashMap<String, double[]> values = new HashMap<String, double[]>();
     double[] values = {intercept, slope};
@@ -344,26 +353,9 @@ public class Calibrator{
     ImagePlus gimg = photo.getGreenChannel();
     ImagePlus bimg = photo.getBlueChannel();
 
-    rimg.show();
+    //rimg.show();
     //gimg.show();
     //bimg.show();
-
-    //RGBPhoto nphoto = null;
-    double max = img.getDisplayRangeMax();
-    double min = img.getDisplayRangeMin();
-    double refvals[] = {0.0,1.0};
-    double ends[] = {min,max};
-    double regressionParams[] = null;
-    double r_Squared = 0.0;
-
-
-    CurveFitter visRegression = new CurveFitter(ends, refvals);
-    visRegression.doFit(0, true);
-    regressionParams = visRegression.getParams();
-    r_Squared = visRegression.getRSquared();
-
-    double intercept = (double)regressionParams[0];
-    double slope = (double)regressionParams[1];
 
     ImagePlus newRedImage = NewImage.createFloatImage((String)"redImage", (int)img.getWidth(), (int)img.getHeight(), (int)1, (int)1);
     ImagePlus newBlueImage = NewImage.createFloatImage((String)"blueImage", (int)img.getWidth(), (int)img.getHeight(), (int)1, (int)1);
@@ -392,36 +384,55 @@ public class Calibrator{
             if( photo.getCameraType().equals(CalibrationPrompt.SURVEY2_RED) ){
               pixel = (double)rimg.getProcessor().getPixelValue(x, y) * calibrationCeofs[1] + calibrationCeofs[0];
 
-              pixel = (int)( (pixel - redReflectMin)/(redReflectMax-redReflectMin)*255 );
-              // histogram stretch map to float-point THIS ONE WORKS!!
-              pixel = (double)(pixel - 0)/(255-0)*1;
+              if( !ReflectanceImageOnly ){
+
+                pixel = (int)( (pixel - redReflectMin)/(redReflectMax-redReflectMin)*255 );
+                // histogram stretch map to float-point THIS ONE WORKS!!
+                pixel = (double)(pixel - 0)/(255-0)*1;
+
+              }
 
               newRedImage.getProcessor().putPixelValue(x, y, pixel);
               newBlueImage.getProcessor().putPixelValue(x, y, 0.0);
               newGreenImage.getProcessor().putPixelValue(x, y, 0.0);
             }else if( photo.getCameraType().equals(CalibrationPrompt.SURVEY2_GREEN) ){
               pixel = (double)gimg.getProcessor().getPixelValue(x, y) * calibrationCeofs[1] + calibrationCeofs[0];
-              pixel = (int)( (pixel - greenReflectMin)/(greenReflectMax-greenReflectMin)*255 );
-              // histogram stretch map to float-point THIS ONE WORKS!!
-              pixel = (double)(pixel - 0)/(255-0)*1;
+
+              if( !ReflectanceImageOnly ){
+
+                pixel = (int)( (pixel - greenReflectMin)/(greenReflectMax-greenReflectMin)*255 );
+                // histogram stretch map to float-point THIS ONE WORKS!!
+                pixel = (double)(pixel - 0)/(255-0)*1;
+
+              }
 
               newRedImage.getProcessor().putPixelValue(x, y, 0.0);
               newGreenImage.getProcessor().putPixelValue(x, y, pixel);
               newBlueImage.getProcessor().putPixelValue(x, y, 0.0);
             }else if( photo.getCameraType().equals(CalibrationPrompt.SURVEY2_BLUE) ){
               pixel = (double)bimg.getProcessor().getPixelValue(x, y) * calibrationCeofs[1] + calibrationCeofs[0];
-              pixel = (int)( (pixel - blueReflectMin)/(blueReflectMax-blueReflectMin)*255 );
-              // histogram stretch map to float-point THIS ONE WORKS!!
-              pixel = (double)(pixel - 0)/(255-0)*1;
+
+              if( !ReflectanceImageOnly ){
+
+                pixel = (int)( (pixel - blueReflectMin)/(blueReflectMax-blueReflectMin)*255 );
+                // histogram stretch map to float-point THIS ONE WORKS!!
+                pixel = (double)(pixel - 0)/(255-0)*1;
+
+              }
 
               newRedImage.getProcessor().putPixelValue(x, y, 0.0);
               newGreenImage.getProcessor().putPixelValue(x, y, 0.0);
               newBlueImage.getProcessor().putPixelValue(x, y, pixel);
             }else if( photo.getCameraType().equals(CalibrationPrompt.SURVEY2_NIR) ){
               pixel = (double)rimg.getProcessor().getPixelValue(x, y) * calibrationCeofs[1] + calibrationCeofs[0];
-              pixel = (int)( (pixel - redReflectMin)/(redReflectMax-redReflectMin)*255 );
-              // histogram stretch map to float-point THIS ONE WORKS!!
-              pixel = (double)(pixel - 0)/(255-0)*1;
+
+              if( !ReflectanceImageOnly ){
+
+                pixel = (int)( (pixel - redReflectMin)/(redReflectMax-redReflectMin)*255 );
+                // histogram stretch map to float-point THIS ONE WORKS!!
+                pixel = (double)(pixel - 0)/(255-0)*1;
+
+              }
 
               newRedImage.getProcessor().putPixelValue(x, y, pixel);
               newBlueImage.getProcessor().putPixelValue(x, y, 0.0);
@@ -439,12 +450,17 @@ public class Calibrator{
     gimg = newGreenImage;
     bimg = newBlueImage;
 
-    rimg.show();
+    //rimg.show();
+
 
     ImagePlus[] nchan = {rimg, gimg, bimg};
 
     RGBPhoto nphoto = new RGBPhoto(nchan, photo.getCameraType(), photo);
     //nphoto.show();
+
+    if( ReflectanceImageOnly ){
+        IJ.save((ImagePlus)nphoto.getImage(), (String)(String.valueOf("C:\\Users\\Peau\\Desktop\\") + nphoto.getFileName() + "_index" + "." + nphoto.getExtension()));
+    }
 
     return nphoto;
   }
@@ -460,14 +476,6 @@ public class Calibrator{
       //gimg.show();
       //bimg.show();
 
-      double max = img.getDisplayRangeMax();
-      double min = img.getDisplayRangeMin();
-      double refvals[] = {0.0,1.0};
-      double ends[] = {min,max};
-      double redRegressionParams[] = null;
-      double blueRegressionParams[] = null;
-      double r_Squared = 0.0;
-
       // @TODO Create the float-point reflectance image
       // img.getBitDepth()
       int imgMax = (int)(Math.pow(2,8)-1);
@@ -479,19 +487,6 @@ public class Calibrator{
       double blueReflectMin = (double)bimg.getDisplayRangeMin() * calibrationCeofs[3] + calibrationCeofs[2];
       double redRange[] = {0.0,redReflectMax};
       double blueRange[] = {0.0,blueReflectMax};
-
-
-      CurveFitter redRegression = new CurveFitter(range, redRange);
-      CurveFitter blueRegression = new CurveFitter(range, blueRange);
-      redRegression.doFit(0, true);
-      blueRegression.doFit(0, true);
-      redRegressionParams = redRegression.getParams();
-      blueRegressionParams = blueRegression.getParams();
-
-      double redIntercept = (double)redRegressionParams[0];
-      double redSlope = (double)redRegressionParams[1];
-      double blueIntercept = (double)blueRegressionParams[0];
-      double blueSlope = (double)blueRegressionParams[1];
 
       double redPixel = 0.0;
       double reflect = 0.0;
@@ -529,73 +524,23 @@ public class Calibrator{
               redPixel = (double)redPixel * calibrationCeofs[1] + calibrationCeofs[0];
               bluePixel = (double)bimg.getProcessor().getPixelValue(x, y) * calibrationCeofs[3] + calibrationCeofs[2];
 
-              //redPixel = redPixel*10;
-              //bluePixel = bluePixel*10;
-
-              //redPixel = redPixel <= 0 ? 0 : redPixel;
-              //greenPixel = 0.0;
-              //bluePixel = bluePixel <= 0 ? 0 : redPixel;
-
-              /*
-              if( redPixel == - bluePixel ){
-                redPixel = 0;
-                bluePixel = 0;
-              }*/
 
               // THIS ONE WORKS!!
-              redPixel = (int)( (redPixel - redReflectMin)/(redReflectMax-redReflectMin)*255 );
-              bluePixel = (int)( (bluePixel - blueReflectMin)/(blueReflectMax-blueReflectMin)*255 );
+
+              if( !ReflectanceImageOnly ){
+                redPixel = (int)( (redPixel - redReflectMin)/(redReflectMax-redReflectMin)*255 );
+                bluePixel = (int)( (bluePixel - blueReflectMin)/(blueReflectMax-blueReflectMin)*255 );
 
 
-              // histogram stretch map to float-point THIS ONE WORKS!!
-              redPixel = (double)(redPixel - 0)/(255-0)*1;
-              bluePixel = (double)(bluePixel - 0)/(255-0)*1;
-
-
-              // Weird attempt
-              //redPixel = (double)(redPixel - 0)/(redReflectMax+blueReflectMax - 0)*1;
-              //bluePixel = (double)(bluePixel - 0)/(blueReflectMax+blueReflectMax - 0)*1;
-
-              /* NaN
-              bluePixel = (double)(bluePixel - 0)/(blueReflectMax-0)*redReflectMax;
-
-              redPixel = (double)(redPixel - 0)/(redReflectMax - 0)*1;
-              bluePixel = (double)(bluePixel = 0)/(blueReflectMax = 0)*1;
-              */
-
-
-              //histogram stretch;
-              //redPixel = (int)( (redPixel - 0)/(255-0)*255 );
-              //bluePixel = (int)( (bluePixel - 0)/(255-0)*255 );
-
-              //IJ.log( (String)"Red reflectance pixel value: " + redPixel );
-
-              // Map back
-              //redPixel = (redPixel - redIntercept)/redSlope;
-              //bluePixel = (bluePixel - blueIntercept)/blueSlope;
-              //redPixel = (redPixel - calibrationCeofs[0])/redSlope;
-              //bluePixel = (bluePixel - calibrationCeofs[2])/blueSlope;
-
-              //redPixel = redPixel*255;
-              //bluePixel = bluePixel*255;
-
-
-              //IJ.log( (String)"New red pixel value: " + redPixel );
-
-              //rimg.getProcessor().putPixelValue(x, y, redPixel);
-              //gimg.getProcessor().putPixelValue(x, y, greenPixel);
-              //bimg.getProcessor().putPixelValue(x, y, bluePixel);
+                // histogram stretch map to float-point THIS ONE WORKS!!
+                redPixel = (double)(redPixel - 0)/(255-0)*1;
+                bluePixel = (double)(bluePixel - 0)/(255-0)*1;
+              }
 
               newRedImage.getProcessor().putPixelValue(x, y, redPixel);
               newBlueImage.getProcessor().putPixelValue(x, y, bluePixel);
               newGreenImage.getProcessor().putPixelValue(x, y, greenPixel);
 
-              /*
-              redPixel = (slope/calibrationCeofs[1])*(double)rimg.getProcessor().getPixelValue(x, y) + (intercept - calibrationCeofs[0])/calibrationCeofs[1];
-              bluePixel = (slope/calibrationCeofs[1])*(double)bimg.getProcessor().getPixelValue(x, y) + (intercept - calibrationCeofs[0])/calibrationCeofs[1];
-              rimg.getProcessor().putPixelValue(x, y, redPixel);
-              bimg.getProcessor().putPixelValue(x, y, bluePixel);
-              */
 
               x++;
           }
@@ -606,11 +551,13 @@ public class Calibrator{
       gimg = newGreenImage;
       bimg = newBlueImage;
 
-      //rimg.show();
-      //bimg.show();
 
       ImagePlus[] nchan = {rimg, gimg, bimg};
       RGBPhoto nphoto = new RGBPhoto(nchan, CalibrationPrompt.SURVEY2_NDVI, photo);
+
+      if( ReflectanceImageOnly ){
+          IJ.save((ImagePlus)nphoto.getImage(), (String)(String.valueOf("C:\\Users\\Peau\\Desktop\\") + nphoto.getFileName() + "_index" + "." + nphoto.getExtension()));
+      }
 
       return nphoto;
   }
@@ -746,18 +693,22 @@ public class Calibrator{
     float targetLength = qr.getTargetSize();
 
     Roi target1Roi = qr.createRectangleRoi( target1Center, 0.3f );
-    //qr.drawPolygonOn( target1Roi, resimg );
+    qr.drawPolygonOn( target1Roi, resimg );
     Roi target2Roi = qr.createRectangleRoi( target2Center, 0.3f );
-    //qr.drawPolygonOn( target2Roi, resimg );
+    qr.drawPolygonOn( target2Roi, resimg );
     Roi target3Roi = qr.createRectangleRoi( target3Center, 0.3f );
-    //qr.drawPolygonOn( target3Roi, resimg );
+    qr.drawPolygonOn( target3Roi, resimg );
 
+    IJ.log("Testing remove me :)");
     target1Roi = qr.mapRoiTo( qrimg, resimg, target1Center, 0.5f );
-    //qr.drawPolygonOn( target1Roi, qrimg );
+    qr.drawPolygonOn( target1Roi, qrimg );
     target2Roi = qr.mapRoiTo( qrimg, resimg, target2Center, 0.5f );
-    //qr.drawPolygonOn( target2Roi, qrimg );
+    qr.drawPolygonOn( target2Roi, qrimg );
     target3Roi = qr.mapRoiTo( qrimg, resimg, target3Center, 0.5f );
-    //qr.drawPolygonOn( target3Roi, qrimg );
+    qr.drawPolygonOn( target3Roi, qrimg );
+
+    resimg.show();
+    qrimg.show();
 
     Roi[] rois = {target1Roi, target2Roi, target3Roi};